pyproject.toml
[project]
name = "nxtwaveragbasedprojects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.6",
    "langchain>=1.2.12",
    "langchain-google-genai>=4.2.1",
    "python-dotenv>=1.2.2",
]
